Only include go-scrap on client binary (no wrapper necessary)
